Определяет логику показа экранов
