+++ js/artist.js
@@ -1,7 +1,9 @@
 import getElementFromTemplate from './get-element-from-template.js';
+import renderScreen from './render-screen.js';
+import genreScreen from './genre.js';
 
 const markup = `
-  <section class="main main--level main--level-genre">
+  <section class="main main--level main--level-artist">
     <a class="play-again play-again__wrap" href="#">
       <img class="play-again__img" src="/img/melody-logo-ginger.png" alt="logo" width="177" height="76">
     </a>
@@ -20,72 +22,61 @@ const markup = `
     <div class="main-mistakes">
       <img class="main-mistake" src="img/wrong-answer.png" width="35" height="49">
       <img class="main-mistake" src="img/wrong-answer.png" width="35" height="49">
-      <img class="main-mistake" src="img/wrong-answer.png" width="35" height="49">
     </div>
 
     <div class="main-wrap">
-      <h2 class="title">Выберите инди-рок треки</h2>
-      <form class="genre">
-        <div class="genre-answer">
-          <div class="player-wrapper">
-            <div class="player">
-              <audio></audio>
-              <button class="player-control player-control--pause"></button>
-              <div class="player-track">
-                <span class="player-status"></span>
-              </div>
-            </div>
+      <h2 class="title main-title">Кто исполняет эту песню?</h2>
+      <div class="player-wrapper">
+        <div class="player">
+          <audio></audio>
+          <button class="player-control player-control--pause"></button>
+          <div class="player-track">
+            <span class="player-status"></span>
           </div>
-          <input type="checkbox" name="answer" value="answer-1" id="a-1">
-          <label class="genre-answer-check" for="a-1"></label>
+        </div>
+      </div>
+      <form class="main-list">
+        <div class="main-answer-wrapper">
+          <input class="main-answer-r" type="radio" id="answer-1" name="answer" value="val-1"/>
+          <label class="main-answer" for="answer-1">
+            <img class="main-answer-preview" src="http://placehold.it/134x134"
+                 alt="Пелагея" width="134" height="134">
+            Пелагея
+          </label>
         </div>
 
-        <div class="genre-answer">
-          <div class="player-wrapper">
-            <div class="player">
-              <audio></audio>
-              <button class="player-control player-control--play"></button>
-              <div class="player-track">
-                <span class="player-status"></span>
-              </div>
-            </div>
-          </div>
-          <input type="checkbox" name="answer" value="answer-1" id="a-2">
-          <label class="genre-answer-check" for="a-2"></label>
+        <div class="main-answer-wrapper">
+          <input class="main-answer-r" type="radio" id="answer-2" name="answer" value="val-2"/>
+          <label class="main-answer" for="answer-2">
+            <img class="main-answer-preview" src="http://placehold.it/134x134"
+                 alt="Краснознаменная дивизия имени моей бабушки" width="134" height="134">
+            Краснознаменная дивизия имени моей бабушки
+          </label>
         </div>
 
-        <div class="genre-answer">
-          <div class="player-wrapper">
-            <div class="player">
-              <audio></audio>
-              <button class="player-control player-control--play"></button>
-              <div class="player-track">
-                <span class="player-status"></span>
-              </div>
-            </div>
-          </div>
-          <input type="checkbox" name="answer" value="answer-1" id="a-3">
-          <label class="genre-answer-check" for="a-3"></label>
+        <div class="main-answer-wrapper">
+          <input class="main-answer-r" type="radio" id="answer-3" name="answer" value="val-3"/>
+          <label class="main-answer" for="answer-3">
+            <img class="main-answer-preview" src="http://placehold.it/134x134"
+                 alt="Lorde" width="134" height="134">
+            Lorde
+          </label>
         </div>
-
-        <div class="genre-answer">
-          <div class="player-wrapper">
-            <div class="player">
-              <audio></audio>
-              <button class="player-control player-control--play"></button>
-              <div class="player-track">
-                <span class="player-status"></span>
-              </div>
-            </div>
-          </div>
-          <input type="checkbox" name="answer" value="answer-1" id="a-4">
-          <label class="genre-answer-check" for="a-4"></label>
-        </div>
-
-        <button class="genre-answer-send" type="submit">Ответить</button>
       </form>
     </div>
   </section>
 `;
 
-export default getElementFromTemplate(markup);
+const artistScreen = getElementFromTemplate(markup);
+
+renderScreen(artistScreen);
+
+const answerList = document.querySelector(`.main-list`);
+
+answerList.addEventListener(`change`, (evt) => {
+  if (evt.target.classList.contains(`main-answer-r`)) {
+    renderScreen(genreScreen);
+  }
+});
+
+export default artistScreen;
