+++ js/render-screen.js
@@ -1,0 +1,6 @@
+const mainEl = document.querySelector(`.main`);
+
+export default (screen) => {
+  mainEl.innerHTML = ``;
+  mainEl.appendChild(screen);
+};
